# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Capto mobile data-collection platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended, disabled]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserStatusRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/language:
    patch:
      operationId: updateMyLanguage
      tags: [users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLanguageRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - name: language
          in: query
          schema:
            type: string
            enum: [en, hi]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    patch:
      operationId: updateCategory
      tags: [categories]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /subcategories:
    get:
      operationId: listSubcategories
      tags: [subcategories]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: categoryId
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: language
          in: query
          schema:
            type: string
            enum: [en, hi]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubcategoryListResponse"
    post:
      operationId: createSubcategory
      tags: [subcategories]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubcategoryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
  /subcategories/{id}:
    get:
      operationId: getSubcategory
      tags: [subcategories]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - name: language
          in: query
          schema:
            type: string
            enum: [en, hi]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
    patch:
      operationId: updateSubcategory
      tags: [subcategories]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubcategoryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
    delete:
      operationId: deleteSubcategory
      tags: [subcategories]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: categoryId
          in: query
          schema:
            type: string
        - name: subcategoryId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, draft]
        - name: collectionType
          in: query
          schema:
            type: string
            enum: [VIDEO, IMAGE, AUDIO]
        - name: language
          in: query
          schema:
            type: string
            enum: [en, hi]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
    post:
      operationId: createTask
      tags: [tasks]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - name: language
          in: query
          schema:
            type: string
            enum: [en, hi]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /tasks/{id}/duplicate:
    post:
      operationId: duplicateTask
      tags: [tasks]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /admin/support-settings:
    get:
      operationId: adminGetSupportSettings
      tags: [support]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportSettings"
    patch:
      operationId: adminUpdateSupportSettings
      tags: [support]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupportSettingsRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportSettings"

  # ─── Admin Static Pages ───────────────────────────────────────────
...
  /public/support:
    get:
      operationId: getPublicSupport
      tags: [public-support]
      parameters:
        - in: query
          name: language
          schema: { type: string, enum: [en, hi] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportSettings"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]
...
    User:
      type: object
      properties:
        id:
          type: string
        phoneNumber:
          type: string
        name:
          type: string
        preferredLanguage:
          type: string
          enum: [en, hi]
          default: en
        status:
          type: string
          enum: [active, inactive, suspended, disabled]
        totalEarnings:
          type: number
        totalSubmissions:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, phoneNumber, status, preferredLanguage, totalEarnings, totalSubmissions, createdAt, updatedAt]
    UpdateLanguageRequest:
      type: object
      properties:
        preferredLanguage:
          type: string
          enum: [en, hi]
      required: [preferredLanguage]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended, disabled]
    UpdateUserStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, inactive, suspended, disabled]
      required: [status]
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]
    CategorySummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
        isActive:
          type: boolean
      required: [id, name, isActive]
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        coverImageUrl:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
        subcategoryCount:
          type: integer
        taskCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, isActive, displayOrder, subcategoryCount, taskCount, createdAt, updatedAt]
...
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        coverImageUrl:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
...
    Subcategory:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/CategorySummary"
        displayOrder:
          type: integer
        isActive:
          type: boolean
        taskCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, categoryId, isActive, displayOrder, taskCount, createdAt, updatedAt]
    CreateSubcategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
      required: [name, categoryId]
    UpdateSubcategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
    SubcategoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subcategory"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        detailedInstructions:
          type: string
        dos:
          type: array
          items:
            type: string
        donts:
          type: array
          items:
            type: string
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/CategorySummary"
        subcategoryId:
          type: string
        subcategory:
          $ref: "#/components/schemas/SubcategorySummary"
        collectionType:
          type: string
          enum: [VIDEO, IMAGE, AUDIO]
        paymentAmount:
          type: number
        currency:
          type: string
        sampleMediaUrl:
          type: string
        minimumDurationSeconds:
          type: integer
        maximumDurationSeconds:
          type: integer
        minimumImageCount:
          type: integer
        maximumImageCount:
          type: integer
        preferredFps:
          type: number
        minimumFps:
          type: number
        preferredCamera:
          type: string
          enum: [REAR, FRONT, ANY]
        preferredLens:
          type: string
          enum: [ULTRA_WIDE, STANDARD, ANY]
        requiredOrientation:
          type: string
          enum: [PORTRAIT, LANDSCAPE, ANY]
        audioRequired:
          type: boolean
        pauseAllowed:
          type: boolean
        maxSubmissionsPerUser:
          type: integer
        maxTotalSubmissions:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        displayOrder:
          type: integer
        status:
          type: string
          enum: [active, inactive, draft]
        submissionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - categoryId
        - collectionType
        - paymentAmount
        - currency
        - preferredCamera
        - preferredLens
        - requiredOrientation
        - audioRequired
        - pauseAllowed
        - displayOrder
        - status
        - submissionCount
        - dos
        - donts
        - createdAt
        - updatedAt
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        detailedInstructions:
          type: string
        dos:
          type: array
          items:
            type: string
        donts:
          type: array
          items:
            type: string
        categoryId:
          type: string
        subcategoryId:
          type: string
        collectionType:
          type: string
          enum: [VIDEO, IMAGE, AUDIO]
        paymentAmount:
          type: number
        currency:
          type: string
        sampleMediaUrl:
          type: string
        minimumDurationSeconds:
          type: integer
        maximumDurationSeconds:
          type: integer
        minimumImageCount:
          type: integer
        maximumImageCount:
          type: integer
        preferredFps:
          type: number
        minimumFps:
          type: number
        preferredCamera:
          type: string
          enum: [REAR, FRONT, ANY]
        preferredLens:
          type: string
          enum: [ULTRA_WIDE, STANDARD, ANY]
        requiredOrientation:
          type: string
          enum: [PORTRAIT, LANDSCAPE, ANY]
        audioRequired:
          type: boolean
        pauseAllowed:
          type: boolean
        maxSubmissionsPerUser:
          type: integer
        maxTotalSubmissions:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        displayOrder:
          type: integer
        status:
          type: string
          enum: [active, inactive, draft]
      required: [title, categoryId]
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        detailedInstructions:
          type: string
        dos:
          type: array
          items:
            type: string
        donts:
          type: array
          items:
            type: string
        categoryId:
          type: string
        subcategoryId:
          type: string
        collectionType:
          type: string
          enum: [VIDEO, IMAGE, AUDIO]
        paymentAmount:
          type: number
        currency:
          type: string
        sampleMediaUrl:
          type: string
        minimumDurationSeconds:
          type: integer
        maximumDurationSeconds:
          type: integer
        minimumImageCount:
          type: integer
        maximumImageCount:
          type: integer
        preferredFps:
          type: number
        minimumFps:
          type: number
        preferredCamera:
          type: string
          enum: [REAR, FRONT, ANY]
        preferredLens:
          type: string
          enum: [ULTRA_WIDE, STANDARD, ANY]
        requiredOrientation:
          type: string
          enum: [PORTRAIT, LANDSCAPE, ANY]
        audioRequired:
          type: boolean
        pauseAllowed:
          type: boolean
        maxSubmissionsPerUser:
          type: integer
        maxTotalSubmissions:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        displayOrder:
          type: integer
        status:
          type: string
          enum: [active, inactive, draft]
    TaskListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]
...
    SupportSettings:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        whatsappNumber:
          type: string
        phoneNumber:
          type: string
          nullable: true
        workingHours:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
      required: [id, email, whatsappNumber, updatedAt]
    UpdateSupportSettingsRequest:
      type: object
      properties:
        email:
          type: string
        whatsappNumber:
          type: string
        phoneNumber:
          type: string
        workingHours:
          type: string
        message:
          type: string
...